requirements.txt
Flask>=0.10.1
leancloud-sdk>=1.0.9
beautifulsoup4>=4.0.0
